+++ sorbet/tapioca/require.rb
@@ -5,4 +5,3 @@
 
 require "ruby_lsp/internal"
 require "ruby_lsp/test_helper"
-require "dotenv"