+++ CHANGELOG.md
@@ -1,24 +1,46 @@
+# Changelog
+
 ## [Unreleased]
+
+- **Breaking change** Remove Dotenv as dependency using native [RubyLSP settings](https://github.com/Shopify/ruby-lsp/pull/2513)
+
 ## [0.4.9] - 2025-01-13
+
 - Updated Ruby LSP version to 0.23.0
+
 ## [0.4.8] - 2024-11-29
+
 - Updated Ruby LSP version to 0.22.0
+
 ## [0.4.7] - 2024-11-18
+
 - Updated Ruby LSP version to 0.21.0 and changed version dependency locking in gemspec to version constraint vía [RubyLSP constraint](https://github.com/Shopify/ruby-lsp/pull/2638) to make it more forward compatible.
+
 ## [0.4.6] - 2024-08-31
+
 - Updated Ruby LSP version to 0.17.17
+
 ## [0.4.5] - 2024-07-12
+
 - Relaxed dotenv version to be more compatible with dotenv older than 3.0
+
 ## [0.4.4] - 2024-07-12
+
 - Updated Ruby LSP version
+
 ## [0.4.3] - 2024-05-27
+
 - Add `dotenv` as a runtime dependency to the gemspec
+
 ## [0.4.0] - 2024-02-10
+
 - Added support to ruby on rails
 - Fixed Class Run on terminal whith wrong pattern
 
 ## [0.3.0] - 2024-02-10
+
 - Removed collisions from different should/context methods that have the same name at different levels
+
 ## [0.2.0] - 2024-02-07
 
 - Add bump gem for semantic versioning and changelog